+++ start/src/App.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
   import { ref } from 'vue'
 
-  const show = ref(true)
+  const show = ref(false)
 </script>
 
 <template>