+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `dialog` 组件新增 `modal-blur` 配置项
 - `dialog` 组件移除 `top` 配置项
 - `dialog` 组件 `width` `height` 配置项支持 `number` 类型
+- 新增 `mask` 组件
 
 **Bug**